scheduled_messages: Move UI to new scheduled_messages_ui module.
