+++ tools/test-js-with-node
@@ -167,6 +167,7 @@ EXEMPT_FILES = make_set(
         "web/src/scheduled_messages.js",
         "web/src/scheduled_messages_overlay_ui.js",
         "web/src/scheduled_messages_popover.js",
+        "web/src/scheduled_messages_ui.js",
         "web/src/scroll_bar.ts",
         "web/src/scroll_util.ts",
         "web/src/search.js",
